+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Support of rocks repository specified in tt config.
+- ``cfg dump`` module. It prints tt environment configuration.
 
 ## [0.3.0] - 2022-12-05